IMC-162 Review Settings
Extracts and enables to run application if different environment and
settings easily, by moving all secrets from environment variables to
encrypted `credentials.yml.enc` file.

To make hierarchical settings easier, and to enable to actually see
which user details is used in which environment, RailsConfig gem is
enabled for the project. This allows easily to have usernames etc as
plain text in the repository while all passwords and API Keys are
encrypted with Rails Encrypted Credentials.


Former-commit-id: 81243701b5398dcc534ac9d86a49a59cbf2bbc3d