Chore: Implement Black-Box Testing
This implements black-box testing with real browsers with click-through
manner. It ensures that our application works with its happy path,
i.e. customers are able to make bookings and things that should always
work, does actually work. This is also underlaying work for implementing
cross-browser testing and visual regression testing in the future.


Former-commit-id: 8a08b95c0185517230925ca9966b75b6f5247fe8