POL-2734 update the users client model to ensure it has one member (#2199)
* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: We need to make the companies gem/engine available in the Users engine, in order to access the ::Companies::Company class from Users::Client. Removing the profiles dependency, also allows us to avoid loops, when bundling gems, for the companies dependency in the users gem.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: Whenever an employee is added to an organization, we want to ensure that their memberships from other companies are removed. We must use Organizations.current_id, instead of client.organization, as calling this will result in a recursive call, to the method "client". The attach_company method was incorrect, as it could have returned nil if the company was not found, instead we want to return the default company. Users::Client needs to have one companies_membership, and a company through a companies_membership, so that multiple memberships for a Users::Client does not ever occur.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
wip

* POL-2734: corrected factories and spec references

* POL-2734: corrected factories and spec references

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: We want the values in companies_memberships.client_id, to be that of member_id, due to the new column client_id that is supposed to replace member_id. A company membership, is also an invalid object thus an update cannot be performed on it, so it is best to destroy it.

* POL-2734 - Update the Users::Client model, to ensure it has one member

POL-2734 - Update the Users::Client model, to ensure it has one member

POL-2734 - Update the Users::Client model, to ensure it has one member

POL-2734 - Update the Users::Client model, to ensure it has one member

Fix: Rubocop fixes.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Revert: Changes no longer needed.

* POL-2734 - Update the Users::Client model, to ensure it has one member

POL-2734 - Update the Users::Client model, to ensure it has one member

Fix: Added a new index, for records where deleted_at is null. This is so that when looking for company memberships quicker, we want to find the ones that have not been soft deleted

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: companies_membership, now uses client, instead of member.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: member type, and member id are not the right columns anymore for companies memberships, instead client_id is.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: When the target is a different object, we need to access the User::Client instance, in different ways. Tests added to test the private method, via its constructor.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: Regenerated the swagger docs with rails docs:api, because of Github actions incompatible API message.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
This spec is no longer needed, as the changes in this branch affect the way the spec is being tested. The current constraints introduced in this branch affect the tests, now make it uneccessary to test this worker, as the migration has already been complete.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
This is no longer needed.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Deleting this, to hopefully resolve merge conflict.

* POL-2734 - Update the Users::Client model, to ensure it has one member

# Assignee: Vahdet Pervan
#
Fix: Regenrated the swagger.json file, as the previous one caused conflicts.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>
Co-authored-by: Warwick Beamish <warwick.beamish@itsmycargo.com>
Co-authored-by: kodiak-itsmycargo[bot] <80279588+kodiak-itsmycargo[bot]@users.noreply.github.com>