POL-2706 customer payment terms (#2209)
* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Feat: We want to add a new column to the Companies company, to stor ethe payment terms details.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Feat: New validation, to populate the payment_terms field if it is an empty string. The strip method removes any blank spaces in the string. Tests added to test the empty string scenario, and the payment_terms value.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Fix: Rubocop fixe.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Fix: It is better that payment_terms is nil, otherwise when this factory is created, it will not be trivial as to why the value would change from "" to nil, due to the validation.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Feat: We want a custom validation instead, as we want to force the client to send the payment_terms param.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Feat: We want to start documenting business rules, so that we can refer to them when needed, instead of the knowledge being retained by individuals.

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Fix: Rubocop errors.

* Update doc/index.rst

Co-authored-by: Mikko Kokkonen <mikko.kokkonen@itsmycargo.com>

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Fix: Much simpler way of validating payment_terms presence

* POL-2706 - Customer Payment Terms

# Assignee: Vahdet Pervan
#
Fix: The payment_terms value may not always be present. Therefore, it needs to be present.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>
Co-authored-by: kodiak-itsmycargo[bot] <80279588+kodiak-itsmycargo[bot]@users.noreply.github.com>
Co-authored-by: Mikko Kokkonen <mikko.kokkonen@itsmycargo.com>