POL-2703 customer payment terms (#2215)
* POL-2703 - PDF: Customer Payment Terms

# Assignee: Vahdet Pervan
#
Style: New style for the payment terms text, on quotation PDF.

* POL-2703 - PDF: Customer Payment Terms

# Assignee: Vahdet Pervan
#
Test: We want to test whether payment_terms contains some info, when it is present on the company instance.

* POL-2703 - PDF: Customer Payment Terms

# Assignee: Vahdet Pervan
#
Feat: New HTML template, for the payment_terms info. The quotation.pdf.erb file, now calls the render_payment_terms to get the payment_terms HTML. In query_decorator, the object being decorated contains the logic to retrieve the company thus the company method is no longer needed.

* POL-2703 - PDF: Customer Payment Terms

# Assignee: Vahdet Pervan
#
Refactor: This test setup does not need to be in the queries factory as it makes a lot of tests fail, thus the factory bot for companies membership only applies in the query decorator test.

* POL-2703 - PDF: Customer Payment Terms

POL-2703 - PDF: Customer Payment Terms

Refactor: Updated the test descriptions, to reflect more on what the test is testing. The companies_membership factory does not need to be created, because it is already available in the spec. Test changed to assert that render_payment_terms is not empty. We want this so that if payment_terms value in companies_company factory changes, this test will still pass.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>
Co-authored-by: kodiak-itsmycargo[bot] <80279588+kodiak-itsmycargo[bot]@users.noreply.github.com>