POL-2705 update customer payment term (#2228)
* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: New route, to access and update an company.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: New controller with two endpoints to retrieve and update a company. Tests added to test the successful response, and response body.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: New serializer for the CompaniesController to use.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Refactor: Forgot to remove this during testing

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: Using the bang operator, will raise an error if there was a problem with updating the company. Without it will just return false, thus a silent error will make it hard to debug what had happened with the update.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Fix: New routes added after running the db migration

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: We want to gracefully handle empty params, by returning a suitable message to the user. Tests for unprocessible entity added too.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
WIP

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Test: New test, to test a 404 scenario for a get request to a company.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Refactor: It is better to reference Companies::Company model in a private method, so that in future if we need to change the model name, we can just change it one place.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Refactor: Snake case params need to be camel cased, to support the front end app. companies_spec has 404 scenario for getting a company. The params structure has now changed, so that all company params live under a company hash.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Refactor: Forgot to camel case this key.

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Feat: Updated swagger doc, for the new endpoints.

* Update engines/api/spec/controllers/api/v2/companies_controller_spec.rb

Co-authored-by: Henry Perschk <henry.perschk@gmail.com>

* POL-2705 - API: Update Customer Payment Term

# Assignee: Vahdet Pervan
#
Refactor: We can use memoization, instead of storing company in a variable.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>
Co-authored-by: kodiak-itsmycargo[bot] <80279588+kodiak-itsmycargo[bot]@users.noreply.github.com>
Co-authored-by: Henry Perschk <henry.perschk@gmail.com>