POL-2830 - Export to email functionality fix (#2293)
* POL-2380 - Journey::Query Load Type Issue

# Assignee: Warwick Beamish
#
Fix: Without .presence, sometimes an empty string would return thus the from header in an email would contain an empty string. @user is defined in the parent class, which is provided through params, and not the query. Tests have been added for offer_email to test the headers, the different values for the from header, and the attachment.

* POL-2830 - Export to Email Function failing

# Assignee: Vahdet Pervan
#
Refactor: Changed the test descriptions, to be more meaningful in regards to what is being tested, alongside the setup.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>