POL-2885 - Add soft delete to legacy local charge model (#2315)
* POL-2855 - Module::DelegationError: signed_id delegated to attachment, but attachment is nil

# Assignee:
#
Feat: We want the legacy charge model to have a deleted_at timestamp, so that we keep the old records under the with_deleted scope. Migration added to achieve this, and the models have been annotated after the migration was ran. Test added, to ensure the legacy local charge instance deleted_at value is present.

* POL-2855 - Module::DelegationError: signed_id delegated to attachment, but attachment is nil

# Assignee:
#
Fix: This annotation had not been committed, when the old route changed.

* POL-2855 - Module::DelegationError: signed_id delegated to attachment, but attachment is nil

# Assignee:
#
Doc: New documentation for the legacy engine, describing mainly what happens when the legacy local charge is deleted.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>