POL-2897 - Return empty hash, when the rate basis cannot be found for a pricing breakdown (#2322)
* POL-2897 - ActionView::Template::Error: undefined method `[]' for nil:NilClass

Fix: Whenever the rate cannot be dug, we want to return an empty hash. If the breakdown's rate_origin type is not a Trucking::Trucking, then we also want to return an empty hash. If the key rate_basis, does not exist within rate, then we also want to return an empty hash, instead of nil.

* POL-2897 - ActionView::Template::Error: undefined method `[]' for nil:NilClass

# Assignee: Vahdet Pervan
#
Refactor: We want the rate_basis method to return either a string, if the rate_basis key can be found in the hash from breakdown_rate_data. Test case missing, for when a rate_origin type is 'Trucking::Trucking'

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>