POL-2896 - Soft delete old notes that belong to a pricing, when adding notes to an existing pricing. PR 2 (#2325)
* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: For any pricing that has outdated notes attached to it, we want to soft delete them, and keep the latest one with the given header (for example Remarks note with the latest date on created_at should be kept). Test added, to ensure that the old notes are soft deleted.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Feat: We want Legacy::Note to have deleted_at, whenever notes are soft deleted.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: Rubocop fixes

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: The range can be used, instead of converting ActiveRecord_Relation to an array and shifting it. Renamed the current_notes variable to be more meaningful.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: Changed the test description, to accurately describe what is happening.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: It is better, to order the deleted_at notes, through updated_at, because in future, the notes may get updated. Tests adjusted for this change to, and we are asserting that the notes that were not soft deleted, contain nil values, and that their body attributes, contain a string.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: Using FactoryBot, to reduce the number of lines of code.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: Disabling this cop, as I do not agree with the example length rule. This test setup and assertions are readable and understandable.

* POL-2896 - Multiple Country Restrictions

Test: New test to verify the latest legacy notes are not soft deleted. Existing test refactored, to only assert on the soft deleted notes.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: Rubocop fix.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: Accidentally put :aggregate_failures inside the string.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Refactor: Moved the tests to a different file, as this file is the sister spec of the pricing inserter. The pricing inserter, now needs to add notes, to a new pricing, instead of an existing one. Whenever a customer uploads a new pricing, in our system, it is a new pricing.

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: We want deleted_at to be indexed

* POL-2896 - Multiple Country Restrictions

# Assignee: Vahdet Pervan
#
Fix: We want deleted_at to be indexed also.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>