POL-2883 clean local charges table in preparation for applications constraints - PR 1 (#2334)
* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Feat: The local charge models now have non unique indexes on the models. The Legacy::LocalCharge model now has the logic to create an upsert id, with the given values.

* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Feat: New worker to remove duplicated local charges. Tests added too.

* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Fix: Migration files, to update the database schema, by removing the unique constraint index, and adding a non unique index.

* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Refactor and fix: Added a guard clause, as the old line was long. Sometimes, counterpart_hub_id can be nil, thus we do not want this, to be included in the SHA1 hash.

* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Refactor: We want counterpart_hub_id to be included in the hasing logic. We also want the generated_upsert_id to be invoked, whenever the system tries to update one of the params listed in the array in generate_upsert_id method.

* POL-2883 - Clean LocalCharges table in preparation for application of constraints

# Assignee: Vahdet Pervan
#
Fix: Bumped the timeout number, as the index addition times out.

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>