POL-2917 shipment request notification (#2352)
* POL-2917 - Shipment Request Notification

# Assignee: Vahdet Pervan
#
Feat: We want to publish an event to the rails event store, when a shipment request is being created.

* POL-2917 - Shipment Request Notification

# Assignee: Vahdet Pervan
#
Feat: We need the engine, to know which notifier to trigger, when the Journey::ShipmentRequestCreated event is triggered.

* POL-2917 - Shipment Request Notification

# Assignee: Vahdet Pervan
#
Feat: This mailer is no longer being used only when an offer is created, it will now be used for when a shipment request is created too, therefore we have removed OfferCreated from the namespaced class, and its references.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: Comment not needed.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feature: This needs to be monetized, so that when commercial value is called from the shipment request, we return a money object for it.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: This factory's client was not the same as the one that belonged to the query, therefore there are two different clients, one that belongs to query, and one that belongs to this factory.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feat: We do not want to return true or false values on the email, because yes or no is more human friendly.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Fix: We want the ShipmentRequestCreatedJob class to be called, whenever we receive a Journey::ShipmentRequestCreated event.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feat: A new notifier job class, to call the admin mailer, for subscribers to the ShipmentRequestCreated event.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feat: Templates for the shipment request created emails.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feat: A new method to send a shipment request created email.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: This has been changed, as ShipmentRequestCreatedJob needs to use this class, but does not use offer.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
The query decorator needs the cargo units decorated, as the shipment request created html file needs dimensions_format from the cargo unit decorator. Quotations.pdf.erb renders html, and in the shipment request created html file, we do not need this html, thus we want to check whether the html is required.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: Changed the styles, and added more checks to check for blank values.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Feat: New method to return meaningful value to the human, if the attribute is not present.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Fix: Rubocop fixes.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Fix: The query decorator now decorates many cargo unts with the cargo unit decorator, therefore the old way was trying to call where and reject on a decorator class, instead of on ActiveRecord.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Fix: Failing tests, now fixed due to changes in the code base.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: Test description updated.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: Renamed the spec file name, to match the class.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: Renamed the subject line, as subject line is being used in multiple places.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: grand_total_section does not need this flag, as the total can be retrieved from the Notifications::ResultDecorator, therefore we do not need to customise this method. Missing tests added too.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: The total can be retrieved now from the total method in Notifications::ResultDecorator, therefore we do not need to check for this value anymore.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Refactor: We want the total to include the currency. The subject lines need to be different for the different methods in the admin mailer. Notifications::OfferSubjectLine is now Notifications::SubjectLine, so find and replace has been done for that. Tests added for the shipment request created mail.

* POL-2917 - Shipment Request Notification / Email

# Assignee: Vahdet Pervan
#
Fix: Fixed failing tests.

* POL-2917: Refactoed Subscription filtering out and fixed specs

* POL-2917: added german templates

* POL-2917: adjusted german templates for boolean.

* POL-2917: added translations for yes and no

* POL-2917: removed miscommited file

* POL-2917: translation for nil

* POL-2917: rubocop

* POL-2917: reset locale after specs

* POL-2917: preview added

* POL-2917: last spec fixed

Co-authored-by: Vahdet Pervan <vahdet.pervan@itsmycargo.com>
Co-authored-by: Warwick Beamish <warwick.beamish@itsmycargo.com>
Co-authored-by: Warwick Beamish <DaddyWozBucks@users.noreply.github.com>